clean __init__.py file
