+++ pytraj/__init__.py
@@ -65,7 +65,10 @@ from .io import (load,
                  write_traj,
                  read_pickle,
                  to_pickle,
+                 select_atoms,
                  )
+
+from .frameiter import iterframe, iterchunk
 
 # alias
 write_trajectory = write_traj
@@ -206,121 +209,6 @@ def set_cpptraj_verbose(cm=True):
 
 set_world_silent(True)
 _verbose = set_cpptraj_verbose
-
-
-def iterframe(traj, *args, **kwd):
-    """create frame iterator with given indices, mask or some iter_options
-
-    Examples
-    --------
-    >>> import pytraj as pt
-    >>> traj = pt.datafiles.load_tz2_ortho()
-    >>> for frame in pt.iterframe(traj, 0, 8, 2): pass
-    >>> for frame in pt.iterframe(traj, 4, mask='@CA'): print(frame)
-    <Frame with 12 atoms>
-    <Frame with 12 atoms>
-    <Frame with 12 atoms>
-    <Frame with 12 atoms>
-    <Frame with 12 atoms>
-    <Frame with 12 atoms>
-
-    # create frame iterator for given indices
-    >>> for frame in pt.iterframe(traj, frame_indices=[0, 7, 3]): print(frame)
-    <Frame with 5293 atoms>
-    <Frame with 5293 atoms>
-    <Frame with 5293 atoms>
-
-
-    >>> fi = pt.iterframe(traj)
-    >>> # iterframe its self
-    >>> fi = pt.iterframe(fi)
-
-    See also
-    --------
-    pytraj.TrajectoryIterator.iterframe
-    """
-    if hasattr(traj, 'iterframe'):
-        return traj.iterframe(*args, **kwd)
-    else:
-        return iterframe_master(traj)
-
-
-def iterchunk(traj, *args, **kwd):
-    """iterate ``traj`` by chunk
-
-    Parameters
-    ----------
-    traj : TrajectoryIterator
-    chunksize : int
-        the number of frames in each chunk
-    start : int, default 0
-        start frame to iterate
-    start : int, default -1 (last frame)
-        stop frame
-    autoimage : bool, default False
-        if True, do autoimage for chunk
-
-    Return
-    ------
-    pytraj.Trajectory, n_frames=chunksize
-        The final chunk might not have the n_frames=chunksize
-
-    Examples
-    --------
-    >>> import pytraj as pt
-    >>> traj = pt.datafiles.load_tz2_ortho()
-    >>> for frame in pt.iterchunk(traj, 4): pass
-    >>> for frame in pt.iterchunk(traj, chunksize=4, start=2): pass
-    >>> for frame in pt.iterchunk(traj, chunksize=4, start=2, stop=9): pass
-    >>> for frame in pt.iterchunk(traj, chunksize=4, autoimage=True): pass
-
-    See also
-    --------
-    pytraj.TrajectoryIterator.iterchunk
-    """
-    return traj.iterchunk(*args, **kwd)
-
-
-def select_atoms(mask, topology):
-    '''return atom indices
-
-    Examples
-    --------
-    >>> import pytraj as pt
-    >>> traj = pt.datafiles.load_tz2_ortho()
-    >>> atom_indices = pt.select_atoms('@CA', traj.top)
-    >>> atom_indices
-    array([  4,  15,  39, ..., 159, 173, 197])
-    >>> pt.select_atoms(traj.top, '@CA')
-    array([  4,  15,  39, ..., 159, 173, 197])
-    '''
-    if isinstance(mask, Topology) and isinstance(topology, string_types):
-        mask, topology = topology, mask
-    return topology.select(mask)
-
-
-select = select_atoms
-
-
-def run(fi):
-    '''shortcut for `for frame in fi: pass`
-
-    >>> import pytraj as pt
-    >>> traj = pt.datafiles.load_tz2_ortho()
-    >>> fi = pt.pipe(traj, ['autoimage', 'rms', 'center :1-13'])
-    >>> pt.run(fi)
-    '''
-    for _ in fi:
-        pass
-
-
-def show():
-    # just delay importing
-    """show plot
-    """
-    from matplotlib import pyplot
-    pyplot.show()
-
 
 def show_versions():
     """
